# api-backend-nodejs/swagger.yaml
openapi: 3.0.3
info:
  title: CutQ Management API
  description: |
    Comprehensive API for CutQ (beauty business) booking and management system.

    ## Features
    - CutQ management with auto-increment display IDs
    - SEO-friendly URLs for better search indexing
    - Service booking and management with verification codes
    - Review and rating system
    - Organized file upload capabilities
    - Special offers and promotions
    - Social authentication (Google, Facebook)
    - Multi-role authentication (Admin/Salon Owner/Customer)

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## SEO-Friendly URLs
    The API supports SEO-optimized URLs for better search engine indexing:
    - CutQ URLs: `/salons/{salon-name}/{display-id}` (e.g., `/salons/beautiful-cuts/123`)
    - Service URLs: `/services/{category}/{service-id}` (e.g., `/services/hair/456`)
    - Backward compatibility maintained for legacy URLs

    ## Display IDs
    The API supports both UUID and auto-increment display IDs for user-friendly URLs:
    - UUIDs: `f12ebd4a-94b0-4162-bd7e-37858ea61d93`
    - Display IDs: `1`, `2`, `3`

  version: 1.2.0
  contact:
    name: CutQ Management API
    email: support@cutq.store
servers:
  - url: http://localhost:3002/api/v1
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message"
...
    Salon:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayId:
          type: integer
          example: 1
          description: "Auto-increment ID for user-friendly URLs"
        name:
          type: string
          example: "Beautiful Cuts"
        description:
          type: string
          example: "Where beauty meets luxury - Professional beauty services"
        address:
          type: string
          example: "123 Beauty Street, Fashion District"
        latitude:
          type: number
          format: float
          nullable: true
          example: 40.7128
        longitude:
          type: number
          format: float
          nullable: true
          example: -74.0060
        phone:
          type: string
          example: "+1-555-0123"
        email:
          type: string
          format: email
          example: "info@beautifulcuts.com"
        rating:
          type: number
          format: float
          example: 4.8
        reviewCount:
          type: integer
          example: 150
        images:
          type: array
          items:
            type: string
          example: ["/uploads/cutqs/beautiful-cuts-1/salon1.jpg", "/uploads/cutqs/beautiful-cuts-1/salon2.jpg"]
          description: "Array of image URLs for auto-scrolling gallery"
        featured:
          type: boolean
          example: true
        isOpen:
          type: boolean
          example: true
        distance:
          type: string
          nullable: true
          example: "2.5 mi"
          description: "Calculated distance from user location"
        specialties:
          type: array
          items:
            type: string
          example: ["Hair", "Nails", "Facial", "Massage"]
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "Parking", "AC", "Refreshments"]
        teamSize:
          type: integer
          example: 8
        yearsInBusiness:
          type: integer
          example: 5
        certifications:
          type: array
          items:
            type: string
          example: ["Licensed Cosmetology", "Organic Certified"]
        workingHours:
          type: object
          example: {
            "monday": {"open": "09:00", "close": "18:00"},
            "tuesday": {"open": "09:00", "close": "18:00"},
            "wednesday": {"open": "09:00", "close": "18:00"},
            "thursday": {"open": "09:00", "close": "20:00"},
            "friday": {"open": "09:00", "close": "20:00"},
            "saturday": {"open": "08:00", "close": "18:00"},
            "sunday": {"open": "10:00", "close": "16:00"}
          }
        slotDuration:
          type: integer
          example: 30
          description: "Appointment slot duration in minutes"
        breakDuration:
          type: integer
          example: 15
          description: "Break duration between appointments in minutes"
        advanceBookingDays:
          type: integer
          example: 30
          description: "How many days in advance bookings are allowed"
        minimumNoticeHours:
          type: integer
          example: 2
          description: "Minimum notice required for bookings in hours"
        enabledPaymentMethods:
          type: array
          items:
            type: string
          example: ["card", "wallet", "cash"]
          description: "Enabled payment methods for this CutQ"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        salonId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Excellent service!"
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /salons:
    get:
      tags:
        - CutQs
      summary: Get all CutQs
      description: |
        Retrieve a list of all CutQs with their details. Supports filtering and pagination
        for better performance and user experience.
      parameters:
        - name: search
          in: query
          description: Search CutQs by name or description
          schema:
            type: string
            example: "hair salon"
        - name: category
          in: query
          description: Filter by service category
          schema:
            type: string
            example: "hair"
        - name: featured
          in: query
          description: Filter by featured status
          schema:
            type: boolean
            example: true
        - name: latitude
          in: query
          description: User's latitude for distance calculation
          schema:
            type: number
            format: float
            example: 40.7128
        - name: longitude
          in: query
          description: User's longitude for distance calculation
          schema:
            type: number
            format: float
            example: -74.0060
        - name: radius
          in: query
          description: Search radius in kilometers
          schema:
            type: number
            format: float
            example: 10
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: CutQs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Salon'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 150
                      pages:
                        type: integer
                        example: 8
                  message:
                    type: string
                    example: "CutQs retrieved successfully"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /salons/{id}:
    get:
      tags:
        - CutQs
      summary: Get CutQ by ID (Legacy)
      description: |
        Retrieve a specific CutQ by ID. Supports both UUID and auto-increment display ID.
        This is the legacy endpoint maintained for backward compatibility.

        Examples:
        - `/salons/1` (display ID)
        - `/salons/856b2dee-18d6-407b-aa9b-1e30af346044` (UUID)
      parameters:
        - name: id
          in: path
          required: true
          description: Salon ID (UUID or display ID)
          schema:
            oneOf:
              - type: string
                format: uuid
                example: "856b2dee-18d6-407b-aa9b-1e30af346044"
              - type: integer
                example: 1
      responses:
        '200':
          description: CutQ retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Salon'
        '404':
          description: CutQ not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Service Endpoints
...
  /reviews/salon/{salonId}:
    get:
      tags:
        - Reviews
      summary: Get reviews for a salon
      description: Retrieve reviews for a specific salon with pagination
      parameters:
        - name: salonId
          in: path
          required: true
          description: Salon ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of reviews per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      reviews:
                        type: array
                        items:
                          $ref: '#/components/schemas/Review'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                            example: 1
                          limit:
                            type: integer
                            example: 10
                          total:
                            type: integer
                            example: 50
                          pages:
                            type: integer
                            example: 5
                      averageRating:
                        type: number
                        example: 4.5
                      totalReviews:
                        type: integer
                        example: 50
...